Refactored the file structure and module names. Removed old tests.
