+++ tests/test_enforce.py
@@ -1,0 +1,12 @@
+import unittest
+
+class EnforceTests(unittest.TestCase):
+    """
+    A container for general tests
+    """
+
+    def test_A(self):
+        self.fail("Not implemented")
+
+if __name__ == '__main__':
+    unittest.main()
