+++ tests/test_validators.py
@@ -1,0 +1,8 @@
+import unittest
+
+class ValidatorsTests(unittest.TestCase):
+    def test_A(self):
+        self.fail("Not implemented")
+
+if __name__ == '__main__':
+    unittest.main()
